replace home page with new /timeline page components
