+++ app/page.tsx
@@ -1,6 +1,7 @@
-import LandingPage from "@/components/LandingPage/LandingPage";
 import { APP_URL } from "@/lib/og/consts";
 import { Metadata } from "next";
+import { TimelineApiProvider as TimelineProvider } from "@/providers/TimelineApiProvider";
+import TimelinePage from "@/components/Timeline/TimelinePage";
 
 export const revalidate = 300;
 
@@ -30,6 +31,12 @@ export const metadata: Metadata = {
   },
 };
 
-const HomePage = () => <LandingPage />;
+const Timeline = () => {
+  return (
+    <TimelineProvider>
+      <TimelinePage />
+    </TimelineProvider>
+  );
+};
 
-export default HomePage;
+export default Timeline;
