--- app/timeline/page.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import { TimelineApiProvider as TimelineProvider } from "@/providers/TimelineApiProvider";
-import TimelinePage from "@/components/Timeline/TimelinePage";
-
-const Timeline = () => {
-  return (
-    <TimelineProvider>
-      <TimelinePage />
-    </TimelineProvider>
-  );
-};
-
-export default Timeline;
